requirements.txt
# Edit requirements.txt and replace the entire file with only the following:

streamlit==1.25.0
yfinance
pandas
numpy
requests
openai
python-dotenv
scikit-learn
tensorflow==2.12.0
alpha_vantage